% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpliedVol.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{Implied Volatility of Option on Future}
\usage{
plotImpliedVol(df)
}
\arguments{
\item{df}{A dataframe containing 6 columns with specific column names: strick, type, optionPrice, future Price, time_to_expiry, interest_rate.}
}
\description{
This function allows you to calculate implied volatility of option on future using Fortran subroutine "zeroin".
}
\details{
At the end, a plot and a dataframe of implied volatility will be returned.
}
\examples{
df <- data.frame(strike = c(50, 20),type = c("C", "P"), optionPrice = c(1.62,0.01),futurePrice = c(48.03, 48.03),time_to_expiry = c(0.1423, 0.1423), interest_rate = c(0.01,0.01))
plotImpliedVol(df)
}

